requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
appdirs==1.4.3            # via black, virtualenv
astroid==2.3.3            # via pylint
attrs==19.3.0             # via black, jsonschema, pytest
backcall==0.1.0           # via ipython
bandit==1.6.2             # via -r requirements-dev.in
black==19.10b0            # via -r requirements-dev.in
bleach==3.1.4             # via nbconvert
blessings==1.7            # via gpustat
certifi==2019.11.28       # via -c requirements.txt, pipenv, requests, sentry-sdk
chardet==3.0.4            # via -c requirements.txt, requests
click==7.1.1              # via -c requirements.txt, black, safety, wandb
configparser==5.0.0       # via wandb
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
distlib==0.3.0            # via virtualenv
docker-pycreds==0.4.0     # via wandb
dparse==0.5.0             # via safety
entrypoints==0.3          # via nbconvert
filelock==3.0.12          # via virtualenv
gevent==1.4.0             # via grequests
gitdb==4.0.2              # via gitpython
gitpython==3.1.0          # via bandit, wandb
gpustat==0.6.0            # via -r requirements-dev.in
gql==0.2.0                # via wandb
gradescope-utils==0.3.1   # via -r requirements-dev.in
graphql-core==1.1         # via gql
greenlet==0.4.15          # via gevent
grequests==0.4.0          # via -r requirements-dev.in
idna==2.9                 # via -c requirements.txt, requests
importlib-metadata==1.6.0  # via jsonschema, pluggy, pytest, virtualenv
ipykernel==5.2.0          # via notebook
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython==7.13.0           # via ipykernel
isort==4.3.21             # via pylint
itermplot==0.331          # via -r requirements-dev.in
jedi==0.16.0              # via ipython
jinja2==2.11.1            # via -c requirements.txt, jupyterlab, jupyterlab-server, nbconvert, notebook
json5==0.9.4              # via jupyterlab-server
jsonschema==3.2.0         # via jupyterlab-server, nbformat
jupyter-client==6.1.2     # via ipykernel, notebook
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook
jupyterlab-server==1.0.7  # via jupyterlab
jupyterlab==2.0.1         # via -r requirements-dev.in
kiwisolver==1.1.0         # via matplotlib
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via -c requirements.txt, jinja2
matplotlib==3.2.1         # via -r requirements-dev.in, itermplot
mccabe==0.6.1             # via pylint
mistune==0.8.4            # via nbconvert
more-itertools==8.2.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r requirements-dev.in
nbconvert==5.6.1          # via notebook
nbformat==5.0.4           # via nbconvert, notebook
nltk==3.4.5               # via -r requirements-dev.in
notebook==6.0.3           # via jupyterlab, jupyterlab-server
numpy==1.18.2             # via -c requirements.txt, itermplot, matplotlib, scipy
nvidia-ml-py3==7.352.0    # via gpustat, wandb
packaging==20.3           # via dparse, pytest, safety
pandocfilters==1.4.2      # via nbconvert
parso==0.6.2              # via jedi
pathspec==0.7.0           # via black
pathtools==0.1.2          # via watchdog
pbr==5.4.4                # via stevedore
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.0.0             # via -r requirements-dev.in
pipenv==2018.11.26        # via dparse
pluggy==0.13.1            # via pytest
prometheus-client==0.7.1  # via notebook
promise==2.3              # via gql, graphql-core
prompt-toolkit==3.0.5     # via ipython
psutil==5.7.0             # via gpustat, wandb
ptyprocess==0.6.0         # via pexpect, terminado
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via -r requirements-dev.in
pydocstyle==5.0.2         # via -r requirements-dev.in
pygments==2.6.1           # via ipython, nbconvert
pylint==2.4.4             # via -r requirements-dev.in
pyparsing==2.4.6          # via matplotlib, packaging
pyrsistent==0.16.0        # via jsonschema
pytest==5.4.1             # via -r requirements-dev.in
python-dateutil==2.8.1    # via jupyter-client, matplotlib, wandb
pyyaml==5.3.1             # via -r requirements-dev.in, bandit, dparse, wandb
pyzmq==19.0.0             # via jupyter-client, notebook
redis==3.4.1              # via redlock-py
redlock-py==1.0.8         # via -r requirements-dev.in
regex==2020.2.20          # via black
requests==2.23.0          # via -c requirements.txt, gql, grequests, safety, wandb
safety==1.8.7             # via -r requirements-dev.in
scipy==1.4.1              # via -c requirements.txt, -r requirements-dev.in
send2trash==1.5.0         # via notebook
sentry-sdk==0.14.3        # via wandb
shortuuid==1.0.1          # via wandb
six==1.14.0               # via -c requirements.txt, astroid, bandit, bleach, blessings, cycler, docker-pycreds, gpustat, gql, graphql-core, itermplot, jsonschema, nltk, packaging, promise, pyrsistent, python-dateutil, stevedore, traitlets, virtualenv, wandb
smmap==3.0.1              # via gitdb
snowballstemmer==2.0.0    # via pydocstyle
stevedore==1.32.0         # via bandit
subprocess32==3.5.4       # via wandb
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
toml==0.10.0              # via -c requirements.txt, black, dparse
tornado==6.0.4            # via -r requirements-dev.in, ipykernel, jupyter-client, jupyterlab, notebook, terminado
traitlets==4.3.3          # via ipykernel, ipython, jupyter-client, jupyter-core, nbconvert, nbformat, notebook
typed-ast==1.4.1          # via astroid, black, mypy
typing-extensions==3.7.4.1  # via mypy
urllib3==1.25.8           # via -c requirements.txt, requests, sentry-sdk
virtualenv-clone==0.5.4   # via pipenv
virtualenv==20.0.15       # via pipenv
wandb==0.8.31             # via -r requirements-dev.in
watchdog==0.10.2          # via wandb
wcwidth==0.1.9            # via prompt-toolkit, pytest
webencodings==0.5.1       # via bleach
wrapt==1.11.2             # via -c requirements.txt, astroid
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
